refactor : 자동차 생성시 검증하는 기능을 도메인에서 구현
