+++ src/test/java/racingcar/utils/CarNameValidatorTest.java
@@ -24,22 +24,4 @@ public class CarNameValidatorTest {
                 }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(ERROR_DUPLICATE_NAME);
     }
-
-    @Test
-    @DisplayName("이름이 5자 초과일 경우")
-    void longName() {
-        Assertions.assertThatThrownBy(() -> {
-                    validateCarName("summer");
-                }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(ERROR_LONG_NAME);
-    }
-
-    @Test
-    @DisplayName("빈 이름이 주어지거나 입력값이 없을 경우")
-    void emptyName() {
-        Assertions.assertThatThrownBy(() -> {
-                    validateCarName("");
-                }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(ERROR_EMPTY_NAME);
-    }
 }
